refactor(ramp): analytics payload getter
